+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next Up
 
+
 - Requires the 5.2.0 or higher of the D&D5e system.
 - A number of internal changes to support the Alchemist Experimental Elixir in item importing, compendium folders and a few other elements.
 - The experimental option to import features on 20th level characters has been disabled due to conflicts with the Munching tool, and breaking features on imported characters.
@@ -7,6 +8,7 @@
 - Alchemist subclass parsing improvements.
 - Subclass proficiencies will now parse correctly for classes imported via Feature Munching.
 - Update default 5e character/monster image in 5.2.
+- Armorer 2024 parsing improvements.
 
 # 6.6.18